+++ angular-project/projects/otec/src/app/app-routing.module.spec.ts
@@ -4,7 +4,6 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { Router } from '@angular/router';
 import {
   HttpClientTestingModule,
-  HttpTestingController,
 } from '@angular/common/http/testing';
 
 import { AppComponent } from './app.component';
@@ -13,7 +12,7 @@ import { routes } from './app-routing.module';
 describe('AppRoutingModule', () => {
   let location: Location;
   let router: Router;
-  let fixture;
+  // let fixture;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
@@ -28,12 +27,11 @@ describe('AppRoutingModule', () => {
     router = TestBed.inject(Router);
     // injecta una instancia de router
     location = TestBed.inject(Location);
-
-    fixture = TestBed.createComponent(AppComponent);
     router.initialNavigation();
   });
 
   it('navegate to "" redirects you to /login', (done) => {
+    localStorage.clear()
     router.navigate(['']).then(() => {
       expect(location.path()).toBe('/login');
       // Llamada asyncronica que ejecuta el expect una vez terminado el router.navigate
@@ -42,16 +40,18 @@ describe('AppRoutingModule', () => {
   });
 
   it('navegate to "login" redirects you to /login', (done) => {
+    localStorage.clear()
     router.navigate(['/login']).then(() => {
       expect(location.path()).toBe('/login');
       done();
     });
   });
 
-  xit('navegate to "dashboard" redirects you to /dashboad', (done) => {
-    // router.navigate(['']).then(() => {
-    //   expect(location.path()).toBe('/login');
-    //   done(); //Llamada asyncronica que ejecuta el expect una vez terminado el router.navigate
-    // });
+  it('navegate to "dashboard" redirects you to /dashboad', (done) => {
+    localStorage.setItem('otec_token', 'testToken');
+    router.navigate(['/dashboard']).then(() => {
+      expect(location.path()).toBe('/dashboard');
+      done(); //Llamada asyncronica que ejecuta el expect una vez terminado el router.navigate
+    });
   });
 });