+++ angular-project/projects/otec/src/app/core/services/auth.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, Optional, Inject } from '@angular/core';
 import * as LoginModel from '../../features/login/login.model';
 import { HttpClient } from '@angular/common/http';
-import { Observable, of } from 'rxjs';
+import { Observable } from 'rxjs';
 
 @Injectable({
   providedIn: 'root',
@@ -14,177 +14,6 @@ export class AuthService {
   ) {
     this.apiBase = environment.api || 'http://localhost:8021';
   }
-
-  // authmock(user: string): Observable<LoginModel.AuthLoginResponse> {
-  //   const arrayMock: { [key: string]: LoginModel.AuthLoginResponse } = {};
-  //   const userlocal = 'jcastill';
-  //   arrayMock[userlocal] = {
-  //     data: {
-  //       token: '0xestasdad',
-  //       nombre_usuario: 'JESSICA LORENA CASTILLO GONZÁLEZ',
-  //       roles_modules: {
-  //         Gestor: {
-  //           id: 2,
-  //           nombre: 'Gestor',
-  //           modulos: {
-  //             Cubicación: {
-  //               nombre: 'Cubicación',
-  //               privilegio: {
-  //                 ver: false,
-  //                 editar: false,
-  //               },
-  //             },
-  //             OT: {
-  //               nombre: 'OT',
-  //               privilegio: {
-  //                 ver: false,
-  //                 editar: false,
-  //               },
-  //             },
-  //           },
-  //         },
-  //       },
-  //     },
-  //     status: {
-  //       responseCode: 0,
-  //       description: 'Consulta OK',
-  //     },
-  //   };
-
-  //   // const userlocal2 = 'carloscj';
-  //   // arrayMock[userlocal2] = {
-  //   //   data: {
-  //   //     token: '0xestasdad',
-  //   //     nombre_usuario: 'Carlos Alberto Campos Jaraquemada',
-  //   //     roles_modules: [
-  //   //       {
-  //   //         id: 5,
-  //   //         nombre: 'Trabajador',
-  //   //         modulos: [
-  //   //           {
-  //   //             id: 1,
-  //   //             nombre: 'OT',
-  //   //           },
-  //   //         ],
-  //   //       },
-  //   //     ],
-  //   //   },
-  //   //   status: {
-  //   //     responseCode: 0,
-  //   //     description: 'Consulta OK',
-  //   //   },
-  //   // };
-
-  //   // const userlocal3 = 'erickuc';
-  //   // arrayMock[userlocal3] = {
-  //   //   data: {
-  //   //     token: '0xestasdad',
-  //   //     nombre_usuario: 'Erick Urrutia Correa',
-  //   //     roles_modules: [
-  //   //       {
-  //   //         id: 3,
-  //   //         nombre: 'Administrador de Contrato',
-  //   //         modulos: [
-  //   //           {
-  //   //             id: 1,
-  //   //             nombre: 'OT',
-  //   //           },
-  //   //         ],
-  //   //       },
-  //   //     ],
-  //   //   },
-  //   //   status: {
-  //   //     responseCode: 0,
-  //   //     description: 'Consulta OK',
-  //   //   },
-  //   // };
-
-  //   // const userlocal4 = 'jaimecc';
-  //   // arrayMock[userlocal4] = {
-  //   //   data: {
-  //   //     token: '0xestasdad',
-  //   //     nombre_usuario: 'Jaime Contreras Cortes',
-  //   //     roles_modules: [
-  //   //       {
-  //   //         id: 4,
-  //   //         nombre: 'Coordinador',
-  //   //         modulos: [
-  //   //           {
-  //   //             id: 1,
-  //   //             nombre: 'OT',
-  //   //           },
-  //   //         ],
-  //   //       },
-  //   //     ],
-  //   //   },
-  //   //   status: {
-  //   //     responseCode: 0,
-  //   //     description: 'Consulta OK',
-  //   //   },
-  //   // };
-
-  //   // const userlocal5 = 'admin';
-  //   // arrayMock[userlocal5] = {
-  //   //   data: {
-  //   //     token: '0xestasdad',
-  //   //     nombre_usuario: 'Admin',
-  //   //     roles_modules: [
-  //   //       {
-  //   //         id: 4,
-  //   //         nombre: 'Administrador OTEC',
-  //   //         modulos: [
-  //   //           {
-  //   //             id: 1,
-  //   //             nombre: 'OT',
-  //   //           },
-  //   //           {
-  //   //             id: 2,
-  //   //             nombre: 'Cubicacion',
-  //   //           },
-  //   //           {
-  //   //             id: 3,
-  //   //             nombre: 'Reportería',
-  //   //           },
-  //   //           {
-  //   //             id: 4,
-  //   //             nombre: 'Administracion',
-  //   //           },
-  //   //         ],
-  //   //       },
-  //   //     ],
-  //   //   },
-  //   //   status: {
-  //   //     responseCode: 0,
-  //   //     description: 'Consulta OK',
-  //   //   },
-  //   // };
-
-  //   // const userlocal6 = 'carlosvb';
-  //   // arrayMock[userlocal6] = {
-  //   //   data: {
-  //   //     token: '0xestasdad',
-  //   //     nombre_usuario: 'Carlos Antonio Valdivia Bustamante',
-  //   //     roles_modules: [
-  //   //       {
-  //   //         id: 6,
-  //   //         nombre: 'Reporteria',
-  //   //         modulos: [
-  //   //           {
-  //   //             id: 3,
-  //   //             nombre: 'Reportería',
-  //   //           },
-  //   //         ],
-  //   //       },
-  //   //     ],
-  //   //   },
-  //   //   status: {
-  //   //     responseCode: 0,
-  //   //     description: 'Consulta OK',
-  //   //   },
-  //   // };
-
-  //   return of(arrayMock[user]);
-  // }
 
   getToken(): string | null {
     return localStorage.getItem('otec_token');
@@ -221,6 +50,25 @@ export class AuthService {
     );
   }
 
+  logOut(user: string, token: string): Observable<LoginModel.LogoutResponse> {
+    console.log('iniciando logout...');
+    const data = {
+      User: user,
+      Token: token,
+    };
+    return (this.httpClient as HttpClient).post<LoginModel.LogoutResponse>(
+      `${this.apiBase}/logout`,
+      JSON.stringify(data)
+    );
+  }
+
+  setUser(user: string): void {
+    localStorage.setItem('user', user);
+  }
+
+  getUser(): string {
+    return localStorage.getItem('user') as string;
+  }
   setPrivilegios(privilegios: object): void {
     const privilegiosJSON = JSON.stringify(privilegios);
     localStorage.setItem('privilegios_user', privilegiosJSON);