+++ Changes.md
@@ -7,6 +7,10 @@
 - Se obtiene, procesa y almacena en localStorage la información del login retornada por el API.
 - Se obtiene la información de login y se despliega en el menú.
 
+### Changes
+
+- Refactoring
+
 ## [0.0.1] - 2021-01-25
 
 ### features