+++ Changes.md
@@ -6,6 +6,7 @@
 
 - Se obtiene, procesa y almacena en localStorage la información del login retornada por el API.
 - Se obtiene la información de login y se despliega en el menú.
+- Se agrega sistema de mensajes de error node-snackbar
 
 ### Changes