+++ Changes.md
@@ -7,10 +7,9 @@
 - Se obtiene, procesa y almacena en localStorage la información del login retornada por el API.
 - Se obtiene la información de login y se despliega en el menú.
 - Se agrega sistema de mensajes de error node-snackbar
+- Se agrega logout con api
 
 ### Changes
-
-- Refactoring
 
 ## [0.0.1] - 2021-01-25